+++ lib/widgets/widgets.dart
@@ -12,6 +12,8 @@ export 'src/clide_button.dart';
 export 'src/clide_card_metrics.dart';
 export 'src/clide_collapser_card.dart';
 export 'src/clide_code_block.dart';
+export 'src/clide_context_menu.dart';
+export 'src/clide_editable.dart';
 export 'src/clide_divider.dart';
 export 'src/clide_settings.dart';
 export 'src/clide_file_image.dart';